+++ packages/atomic/src/components/facets/facet-value-link/facet-value-link.tsx
@@ -14,14 +14,15 @@ export const FacetValueLink: FunctionalComponent<FacetValueProps> = (
   });
 
   return (
-    <li key={props.displayValue}>
+    <li key={props.displayValue} class={props.class}>
       <Button
         style="text-neutral"
-        part="value-link"
+        part={props.part ?? 'value-link'}
         onClick={() => props.onClick()}
         class="group w-full flex items-center px-2 py-2.5 text-left truncate no-outline"
         ariaPressed={props.isSelected.toString()}
         ariaLabel={ariaLabel}
+        ref={props.buttonRef}
       >
         {children}
         <span part="value-count" class="value-count">